pyproject.toml
[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "deepiri-zepgpu"
version = "0.1.0"
description = "Serverless GPU framework for kernel-as-a-service computing"
authors = ["Deepiri Team <team@deepiri.ai>"]
readme = "README.md"
packages = [{ include = "deepiri_zepgpu" }]
keywords = ["gpu", "serverless", "cuda", "deep-learning", "simulation", "kernel"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Physics",
    "Topic :: System :: Distributed Computing",
]

[tool.poetry.dependencies]
python = "^3.11"
click = ">=8.1.0"
typer = ">=0.9.0"
rich = ">=13.7.0"
httpx = ">=0.26.0"
tenacity = ">=8.2.0"
fastapi = ">=0.109.0"
uvicorn = { extras = ["standard"], version = ">=0.27.0" }
pydantic = { extras = ["email"], version = ">=2.6.0" }
pydantic-settings = ">=2.1.0"
python-jose = { extras = ["cryptography"], version = ">=3.3.0" }
pyjwt = ">=2.8.0"
passlib = { extras = ["bcrypt"], version = ">=1.7.4" }
python-multipart = ">=0.0.9"
websockets = ">=12.0"
sqlalchemy = { extras = ["asyncio"], version = ">=2.0.0" }
asyncpg = ">=0.29.0"
alembic = ">=1.13.0"
redis = ">=5.0.0"
celery = { extras = ["redis"], version = ">=5.3.0" }
boto3 = ">=1.34.0"
aiobotocore = ">=2.9.0"
prometheus-client = ">=0.19.0"
structlog = ">=24.1.0"
pynvml = ">=11.5.0"
cloudpickle = ">=3.0.0"
psutil = ">=5.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.2.0"
mypy = "^1.8.0"

[tool.poetry.group.gpu]
optional = true

[tool.poetry.group.gpu.dependencies]
torch = ">=2.2.0"
torchvision = ">=0.17.0"
cupy-cuda12x = ">=13.0.0"
numba = ">=0.59.0"

[tool.poetry.group.ml]
optional = true

[tool.poetry.group.ml.dependencies]
torch = ">=2.2.0"
jax = ">=0.4.25"
optax = ">=0.2.0"
flax = ">=0.8.0"
pytorch-lightning = ">=2.2.0"

[tool.poetry.scripts]
deepiri-gpu = "deepiri_zepgpu.cli:main"

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.pytest_cache
  | \.ruff_cache
  | __pycache__
  | build
  | dist
  | \.venv
)/
'''

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # Pyflakes
    "I",   # isort
    "C",   # flake8-comprehensions
    "B",   # flake8-bugbear
    "UP",  # pyupgrade
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["deepiri_zepgpu"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/site-packages/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]
